+++ src/main/java/net/jasper/mod/gui/option/PlayerAutomaOptionsScreen.java
@@ -176,6 +176,15 @@ public class PlayerAutomaOptionsScreen extends GameOptionsScreen {
             (bool) -> (bool ? ScreenTexts.ON : ScreenTexts.OFF)
     );
 
+    public static OptionButton<Boolean> stopReplayOnManualInput = new OptionButton<>(
+            true,
+            OptionButton.BOOLEAN_VALUES,
+            "playerautoma.option.stopReplayOnManualInput",
+            Object::toString,
+            Boolean::parseBoolean,
+            (bool) -> (bool ? ScreenTexts.ON : ScreenTexts.OFF)
+    );
+
 
     public static ButtonWidget openCommandsToExclude = ButtonWidget.builder(
             Text.translatable("playerautoma.option.openCommandsToExclude"),
@@ -280,6 +289,9 @@ public class PlayerAutomaOptionsScreen extends GameOptionsScreen {
         showQuickSlotsInQuickMenuButton.setTooltip(Tooltip.of(Text.translatable("playerautoma.option.tooltip.showQuickSlotsInQuickMenu")));
         showQuickSlotsInQuickMenu.setButton(showQuickSlotsInQuickMenuButton);
 
+        ButtonWidget stopReplayOnManualInputButton = stopReplayOnManualInput.buttonOf();
+        stopReplayOnManualInputButton.setTooltip(Tooltip.of(Text.translatable("playerautoma.option.tooltip.stopReplayOnManualInput")));
+        stopReplayOnManualInput.setButton(stopReplayOnManualInputButton);
 
         ButtonWidget openKeyBindOptionsButton = ButtonWidget.builder(
                 Text.translatable("playerautoma.option.openKeyBindings"),
@@ -299,12 +311,13 @@ public class PlayerAutomaOptionsScreen extends GameOptionsScreen {
         adder.add(recordInventoryActivitiesButton);
         adder.add(alwaysPreventMenuButton);
         adder.add(resetKeyBindingsOnRecordingButton);
+        adder.add(showQuickSlotsInQuickMenuButton);
         adder.add(EmptyWidget.ofHeight(4), 2);
 
         adder.add(useCTRLForQuickSlotsButton);
         adder.add(useALTForQuickSlotsButton);
         adder.add(preventSlotChangesButton);
-        adder.add(showQuickSlotsInQuickMenuButton);
+        adder.add(stopReplayOnManualInputButton);
         adder.add(EmptyWidget.ofHeight(4), 2);
 
         adder.add(recordCommands.button);